Add features to draw, pick board size & draw style
Create a prompt that triggers when the user uses the "Create new board"
button. Allowing the user to set up the size of it's drawing board, the
input has to be a number not higher than 100. Otherwise it alerts the
user and ask for its input again. The button to create the board, also
resets the board if there's one already layout. It also resets all the
buttons state.

Create a drawing function to draw on the board. Create a Pencil button
to add opacity sensitivity for new inputs from the drawing tool and
create a colorful button to get random colors.